# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: QueueCutter API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forms:
    get:
      operationId: listForms
      tags: [forms]
      summary: List available forms
      description: Returns the catalog of supported government forms, optionally filtered by country
      parameters:
        - name: countryCode
          in: query
          required: false
          schema:
            type: string
          description: Filter forms by country code (US, IN, GB)
      responses:
        "200":
          description: List of forms
          content:
            application/json:
              schema:
                type: object
                properties:
                  forms:
                    type: array
                    items:
                      $ref: "#/components/schemas/FormSummary"
                required: [forms]
  /forms/{formId}:
    get:
      operationId: getForm
      tags: [forms]
      summary: Get form details
      description: Returns full form schema including questions and field mappings
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDetail"
        "404":
          description: Form not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    FormSummary:
      type: object
      properties:
        id:
          type: string
        countryCode:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        whoItIsFor:
          type: string
        estimatedMinutes:
          type: integer
        questionCount:
          type: integer
        category:
          type: string
      required: [id, countryCode, name, shortDescription, whoItIsFor, estimatedMinutes, questionCount, category]
    FormQuestion:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        hint:
          type: string
        hintHi:
          type: string
        type:
          type: string
          enum: [text, date, select, radio, number, textarea, yesno]
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
        fieldMapping:
          type: string
        officialLabel:
          type: string
        validationPattern:
          type: string
        conditionalOn:
          type: object
          properties:
            questionId:
              type: string
            value:
              type: string
          required: [questionId, value]
      required: [id, text, type, required, fieldMapping, officialLabel]
    FormDetail:
      type: object
      properties:
        id:
          type: string
        countryCode:
          type: string
        name:
          type: string
        officialName:
          type: string
        shortDescription:
          type: string
        longDescription:
          type: string
        whoItIsFor:
          type: string
        category:
          type: string
        requiredDocuments:
          type: array
          items:
            type: string
        commonRejectionReasons:
          type: array
          items:
            type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/FormQuestion"
        submissionOffice:
          type: string
        submissionMethod:
          type: string
        processingTime:
          type: string
        fee:
          type: string
        disclaimer:
          type: string
      required: [id, countryCode, name, officialName, shortDescription, longDescription, whoItIsFor, category, requiredDocuments, commonRejectionReasons, questions, submissionOffice, submissionMethod, processingTime, fee, disclaimer]
...